Move BreakInIndexes to utils
